split up AuthDomainSpec test suite
